+++ inst/sql/sql_server/GetCohorts.sql
@@ -1,15 +1,10 @@
-{DEFAULT @cdm_version = '5'}
 {DEFAULT @target_id = ''}
 {DEFAULT @sampled = FALSE}
 
 SELECT row_id,
 	person_seq_id,
 	CAST(cohort.subject_id AS VARCHAR(30)) AS person_id,
-{@cdm_version == "4"} ? {	
-	CASE WHEN cohort_concept_id = @target_id THEN 1 ELSE 0 END AS treatment,
-} : {
 	CASE WHEN cohort_definition_id = @target_id THEN 1 ELSE 0 END AS treatment,
-}
 	cohort_start_date,
 	days_from_obs_start,
 	days_to_cohort_end,