+++ inst/sql/sql_server/GetOutcomes.sql
@@ -22,7 +22,6 @@ limitations under the License.
 {DEFAULT @outcome_database_schema = 'CDM_SIM' } 
 {DEFAULT @outcome_table = 'condition_era' }
 {DEFAULT @outcome_ids = ''} 
-{DEFAULT @cdm_version = '5'}
 {DEFAULT @sampled = FALSE}
 
 {@outcome_table == 'condition_era' } ? {
@@ -47,11 +46,7 @@ WHERE DATEDIFF(DAY, condition_era_start_date, cohort_start_date) <= days_from_ob
 	AND DATEDIFF(DAY, cohort_start_date, condition_era_start_date) <= days_to_obs_end
 } : {
 SELECT DISTINCT row_id,
-{@cdm_version == "4"} ? {	
-	outcome.cohort_concept_id AS outcome_id,
-} : {
 	outcome.cohort_definition_id AS outcome_id,
-}	
 	DATEDIFF(DAY, cohort_person.cohort_start_date, outcome.cohort_start_date) AS days_to_event
 {@sampled} ? {
 FROM #cohort_sample cohort_person
@@ -62,9 +57,5 @@ INNER JOIN @outcome_database_schema.@outcome_table outcome
 	ON cohort_person.subject_id = outcome.subject_id
 WHERE DATEDIFF(DAY, outcome.cohort_start_date, cohort_person.cohort_start_date) <= days_from_obs_start
 	AND DATEDIFF(DAY, cohort_person.cohort_start_date, outcome.cohort_start_date) <= days_to_obs_end
-{@cdm_version == "4"} ? {	
-	AND outcome.cohort_concept_id IN (@outcome_ids)
-} : {
 	AND outcome.cohort_definition_id IN (@outcome_ids)
-}		
 }